requirements.txt
flask==3.0.0
python-dotenv==1.0.0
plotly==5.18.0
pandas==2.1.3
numpy==1.26.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
vanna==0.0.98
werkzeug==3.0.1